requirements.txt
aliyun-python-sdk-core==2.13.26
aliyun-python-sdk-pvtz==1.0.4
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
cryptography==3.2.1
idna==2.10
jmespath==0.10.0
loguru==0.5.3
ntlm-auth==1.5.0
pycparser==2.20
pycryptodome==3.9.9
pywinrm==0.4.1
requests==2.24.0
requests-ntlm==1.1.0
six==1.15.0
urllib3==1.25.11
xmltodict==0.12.0
